Issue #69: Move UI elements to separate file.
